requirements.txt
streamlit~=1.17.0
setuptools~=63.2.0
requests~=2.28.1
python-dotenv~=0.21.0